% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{basic_linear_regression}
\alias{basic_linear_regression}
\title{Function performs lm over accross a number of genes}
\usage{
basic_linear_regression(expr_mat, p_data, model_formula)
}
\arguments{
\item{expr_mat}{matrix of expression values}

\item{p_data}{matrix with sample data}

\item{model_formula}{a string of the form: y ~ x + o}
}
\value{
lm list
}
\description{
Function performs lm over accross a number of genes
}
