% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{basic_logistic_regression}
\alias{basic_logistic_regression}
\title{Performs logistic regression with numerous Y vals}
\usage{
basic_logistic_regression(expr_mat, p_data, model_formula)
}
\arguments{
\item{expr_mat}{matrix of expression values}

\item{p_data}{matrix with sample data}

\item{model_formula}{a string of the form: y ~ x + o}
}
\value{
lm list
}
\description{
Performs logistic regression with numerous Y vals
}
