% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{better_model_matrix}
\alias{better_model_matrix}
\title{Creates a much nicer formatted design matrix than the default model.matrix}
\usage{
better_model_matrix(..., data, show_warnings = TRUE)
}
\arguments{
\item{...}{formula with outcome as first variable}

\item{data}{data.frame}

\item{show_warnings}{logical whether to show wanrings}
}
\value{
model.matrix
}
\description{
I made this function because the way the model.matrix function
creates a desing matrix, can cause issues with differential
gene expression tools, such as edgeR, DESeq2,
and limma. If the intercept column is included,
it first removes the parens and parens from the
intercept column and then removing the variable
name from the colnames and replaces them with
factor level vs reference level syntax. If the
design is specified without an intercept, variable
concatentation is simply removed.
}
