% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{clean_by_levels}
\alias{clean_by_levels}
\title{Removes un-informative cols from a data.frame}
\usage{
clean_by_levels(dat, n_levels = 1)
}
\arguments{
\item{dat}{a data.frame}

\item{n_levels}{thresholded number of items to keep column}
}
\value{
a data.frame without un-informative columns
}
\description{
Removes un-informative cols from a data.frame
}
\examples{
library(delutils)
dat <- data.frame(A=rep(1,4), B=seq(1,4), C=c('f1', 'f2', 'f1', 'f1'))
cleaned_dat <- clean_by_levels(dat)
}
