% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{color_dna_sequence}
\alias{color_dna_sequence}
\title{Colors a vector of DNA/peptide sequences via html}
\usage{
color_dna_sequence(sequences)
}
\arguments{
\item{sequences}{vector of DNA or peptide sequences}
}
\description{
Colors a vector of DNA/peptide sequences via html
}
\examples{
#' library(knitr)
library(kableExtra)

# Example 1: DNA sequences
dna_sequences <- c("ACTG", "GCTA", "TACG")
colored_dna_sequences <- color_dna_sequence(dna_sequences)

# Creating a table with kable for DNA sequences
dna_data <- data.frame(
  DNA_Sequence = dna_sequences,
  Colored_DNA_Sequence = colored_dna_sequences
)
dna_data \%>\%
  kable('html', escape = FALSE) \%>\%
  kable_styling('striped', full_width = FALSE)


# Example 2: Peptide sequences
peptide_sequences <- c("MKTAY", "LLGTP", "IVAGL")
colored_peptide_sequences <- color_dna_sequence(peptide_sequences)

# Creating a table with kable for peptide sequences
peptide_data <- data.frame(
  Peptide_Sequence = peptide_sequences,
  Colored_Peptide_Sequence = colored_peptide_sequences
)
peptide_data \%>\%
  kable('html', escape = FALSE) \%>\%
  kable_styling('striped', full_width = FALSE)

}
