% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{de_rbind_df}
\alias{de_rbind_df}
\title{de_rbind_df}
\usage{
de_rbind_df(df, col_to_split = 1)
}
\arguments{
\item{df}{data.frame}

\item{col_to_split}{columnt to split data.frame by}
}
\value{

}
\description{
de_rbind_df
}
\keyword{internal}
