% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_wrappers.R
\name{dge_wrapper}
\alias{dge_wrapper}
\title{Wrapper function to create dge from txi}
\usage{
dge_wrapper(txi, p, gene_annots)
}
\arguments{
\item{txi}{'lengthScaledTPM' txi object}

\item{p}{data.frame of sample phenodata}

\item{gene_annots}{gene annotations}
}
\value{
DGE object
}
\description{
Wrapper which returns a dge that has been filtered
and is ready for downstream processing. I've
added a bunch of errors here which should prevent
mistakes in creating/using this object
}
