% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{.create_test_dds}
\alias{.create_test_dds}
\title{Title}
\usage{
.create_test_dds(n_genes = 500, m_samples = 60, seed = 42, groups = NULL)
}
\arguments{
\item{groups}{}
}
\description{
Title
}
\keyword{internal}
