% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{.lmlist_to_df}
\alias{.lmlist_to_df}
\title{Turn summary object from lm into a data.frame}
\usage{
.lmlist_to_df(lmlist, coef = 2)
}
\arguments{
\item{lmlist}{lm summary object}

\item{coef}{the coefficient to extract}
}
\value{
data.frame with coefficient
}
\description{
Turn summary object from lm into a data.frame
}
\keyword{internal}
