% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{.split_by_reg}
\alias{.split_by_reg}
\title{Split a named vector of DEGs by fold change}
\usage{
.split_by_reg(lfc_list)
}
\arguments{
\item{lfc_list}{DEG vector of FCs}
}
\description{
Split a named vector of DEGs by fold change
}
\keyword{internal}
