% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_wrappers.R
\name{enrich_wrapper}
\alias{enrich_wrapper}
\title{Over Representation analysis (ORA)}
\usage{
enrich_wrapper(res_object, lfc_thresh = 0)
}
\arguments{
\item{res_object}{res object in long(ish) format}

\item{lfc_thresh}{lfc threshold  absolute value}
}
\value{
data.frame of enriched pathways
}
\description{
ORA wrapper which splits the long res list into
combined, up-regulated, and down-regulated list of
genes. It then performs ORA using clusterProfiler
}
