% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{expression_summaries}
\alias{expression_summaries}
\title{Returns summaries of gene expression}
\usage{
expression_summaries(
  long_res_table,
  log2_thresh = 0,
  logFC_col = NULL,
  padj_col = NULL
)
}
\arguments{
\item{long_res_table}{res df in long format}

\item{log2_thresh}{threshold to genes by lfc}

\item{logFC_col}{str of logFC_col}

\item{padj_col}{str of padj_col}
}
\value{
table summary
}
\description{
This just summarizes the number of significant up & down
regulated genes per group
}
