% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{find_inf_rows}
\alias{find_inf_rows}
\title{Finds rows in a data.frame which contain Inf values}
\usage{
find_inf_rows(df)
}
\arguments{
\item{df}{data.frame or tibble}
}
\value{
A vector of unique row indices containing Inf values
}
\description{
Finds rows in a data.frame which contain Inf values
}
\examples{
# Example usage:
dat <- data.frame(A = c(Inf, 2:4), B = c(1:3, Inf), C = 1:4)
find_inf_rows(dat)
}
