% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{find_na_rows}
\alias{find_na_rows}
\title{Finds rows in a data.frame which contain NA values}
\usage{
find_na_rows(df)
}
\arguments{
\item{df}{data.frame or tibble}
}
\value{
indexes of rows with NA values
}
\description{
Finds rows in a data.frame which contain NA values
}
\examples{
library(delutils)
dat <- data.frame(A=c(NA,2:4), B=c(1:3, NA), C=1:4)
find_na_rows(dat)

}
