% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{gene_quantile_filter}
\alias{gene_quantile_filter}
\title{Gene expression filter by quantile}
\usage{
gene_quantile_filter(expr_mat, quant_val = 0.05)
}
\arguments{
\item{expr_mat}{an expression matrix of class matrix}

\item{quant_val}{quantile threshold}
}
\value{
filtered by quantile expresion matrix
}
\description{
This function first removes rows in which
the max value equals the min value and
then filters the resulting genes by row
medians. Rows with a median greater than
the resulting value of the quantile are
retained.
}
