% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{get_deseq2_results}
\alias{get_deseq2_results}
\title{Returns DESeq2 results in long format}
\usage{
get_deseq2_results(dds_object, gene_annots = NULL, use_shrinkage = FALSE)
}
\arguments{
\item{dds_object}{DDS object after running DESeq}

\item{gene_annots}{Gene annotations which if not given are assumed}

\item{use_shrinkage}{logical whether to perform lfcshrinkage}
}
\value{
data.frame of results in long(ish) format
}
\description{
Returns DESeq2 results in long format
}
