% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_geneSets}
\alias{get_geneSets}
\title{Extract genes sets from a pathway results data.frame}
\usage{
get_geneSets(path_df, delim = "/")
}
\arguments{
\item{path_df}{data.frame with gene sets}

\item{delim}{delimeter seperating genes (default: '/')}
}
\value{
list with names as pathway ids
}
\description{
pathway results data.frame with a column
of gene sets seperated by a delimiter
}
