% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_lfc}
\alias{get_lfc}
\title{Gets lfc from a pathway res df}
\usage{
get_lfc(gene_list, res_df)
}
\arguments{
\item{gene_list}{list of gene vectors}

\item{res_df}{long res df table}
}
\value{

}
\description{
This returns the lfc from a genelist as a
named vector of genes by lfc
}
\keyword{internal}
