% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{get_overlap_matix}
\alias{get_overlap_matix}
\title{Creates matrix of overlap ratios}
\usage{
get_overlap_matix(l1, l2 = NULL)
}
\arguments{
\item{l1}{gene list}

\item{l2}{optional 2nd gene list}
}
\value{
top diagonal matrix of overlap ratios
}
\description{
This function creates a matrix of the ratio
of overlapping genesets. It takes a list of genes
or alternatively 2 gene lists
}
\examples{
\dontrun{
# returns a self (l1 vs l1) overlap matrix
genes <- paste0('gene', seq(1, 100))
set.seed(42)
n_genes_to_sample <- rnbinom(10, 10, .5)
l1 <- lapply(seq_along(n_genes_to_sample), function(i){
 n <- n_genes_to_sample[i]
 sample(genes, n)
}) \%>\%
 setNames(., paste0('L1_', LETTERS[1:10]))
omat <- get_overlap_matix(l1)
}

}
