% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{increase_xy_lims}
\alias{increase_xy_lims}
\title{Returns expanded coordinates of a ggplot}
\usage{
increase_xy_lims(plt, prop = 0.25)
}
\arguments{
\item{plt}{ggplot object}

\item{prop}{proportion to exand the x & y limits}
}
\value{
list with x & y limits as vector
}
\description{
This returns coordinates which will expand
the plotting area when used with
}
\examples{
\dontrun{
xy_lims <- increase_xy_lims(plt, .15)
plt + coord_cartesian(xlim = xy_lims$x_lims, ylim = xy_lims$y_lims)
}

}
