% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{jaccard}
\alias{jaccard}
\title{Calculates the jaccard distance between 2 named sets}
\usage{
jaccard(x, y, check_sets = TRUE)
}
\arguments{
\item{x}{set vector}

\item{y}{set vector}
}
\value{
jaccard distance
}
\description{
Calculates the jaccard distance between 2 named sets
}
