% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{make_flextable}
\alias{make_flextable}
\title{Makes a flextable that can be (sometimes, atleast) used as-is}
\usage{
make_flextable(dat_tbl)
}
\arguments{
\item{dat_tbl}{data.frame or tibble to be turned into a flextable}
}
\value{
flextable object
}
\description{
This function creates a flextable using customizations that I generally like.
It returns a flextable object which can be further customized.
}
