% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{order_by_similarity}
\alias{order_by_similarity}
\title{Order DNA sequences by Levenshtein edit distance}
\usage{
order_by_similarity(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector containing DNA sequences (A, C, G, T).}
}
\value{
A character vector of DNA sequences ordered by similarity.
}
\description{
This function takes a vector of DNA sequences, computes pairwise distances
between them, and orders the sequences based on their similarity using
hierarchical clustering.
}
\examples{
dna_strings <- c("ACTG", "ACCG", "ATGC", "AGTC")
ordered_dna <- order_by_similarity(dna_strings)
print(ordered_dna)

}
