% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{overlap_ratio}
\alias{overlap_ratio}
\title{Calculates the overlap coefficient of 2  sets}
\usage{
overlap_ratio(x, y, check_sets = TRUE)
}
\arguments{
\item{x}{set vector}

\item{y}{set vector}
}
\value{
overlap ratio/coefficient
}
\description{
Calculates the overlap coefficient of 2  sets
}
