% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_named_df_list.R
\name{rbind_named_df_list}
\alias{rbind_named_df_list}
\title{rbinds a named list of data.frames}
\usage{
rbind_named_df_list(df_list, sum_col_name = "col_from_list")
}
\arguments{
\item{df_list}{named list of data.frames}

\item{sum_col_name}{colname of new column bound to data.frame}
}
\value{
a single data.frame row bound data.frame
}
\description{
rbinds together data.frames in which the list name
is turned into a column and bound to the data.frame
becoming the first column of the data.frame
}
\examples{
library(dplyr)
library(delutils)
l <- list(data.frame(A=seq(1,5)), data.frame(A=seq(6, 7))) \%>\%
 setNames(., c('Cond1', 'Cond2'))
 rbind_named_df_list(l)

# Doesn't work because list contains matrix
\dontrun{
l <- list(data.frame(A=seq(1,5)),
 data.frame(A=seq(6, 7)),
 matrix(data=NA, ncol=1, nrow=5)) \%>\%
 setNames(., c('Cond1', 'Cond2', 'Cond3'))
 rbind_named_df_list(l)
}
}
