% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_utils.R
\name{select_genes_by_variance}
\alias{select_genes_by_variance}
\title{Selects the top n genes by variance}
\usage{
select_genes_by_variance(expr_mat, n_genes = 5000)
}
\arguments{
\item{expr_mat}{expression mat}

\item{n_genes}{number of genes to extract}
}
\value{
gene matrix
}
\description{
Selects the top n genes by variance
}
