% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{split_fix_names}
\alias{split_fix_names}
\title{Split names by delim character}
\usage{
split_fix_names(df, delim = ":", split_item = 1, remove_space = TRUE)
}
\arguments{
\item{df}{a data.frame}

\item{delim}{character to split names by}

\item{split_item}{index of split string to keep}

\item{remove_space}{should whitespace be replaced with "_" default: TRUE}
}
\value{
vector of names
}
\description{
Split names by delim character
}
\keyword{internal}
