% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_helpers.R
\name{theme_large_font}
\alias{theme_large_font}
\title{Increases ggplot2 theme size}
\usage{
theme_large_font(rel_size = 1.5)
}
\arguments{
\item{rel_size}{}
}
\description{
Functions which increases the relative size of
fonts used in ggplot2. It also makes some other
adjustments like increasing margin size, ticks,
etc.
}
\keyword{internal}
