% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{uniquefy_base}
\alias{uniquefy_base}
\title{Base function which does the uniquefying}
\usage{
uniquefy_base(dat, group_col, val_col)
}
\arguments{
\item{dat}{data frame}

\item{group_col}{group_col string denoting group column (col with duplicates)}

\item{val_col}{string denoting value column}
}
\value{
logical vector
}
\description{
This is meant to be an internal function
which is the commonly used across the
different uniquefiers. It works by
taking uniquefying based on largest value
and returning a logical vector
}
\keyword{internal}
