% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{uniquefy_by_variance}
\alias{uniquefy_by_variance}
\title{Uniquefy by variance}
\usage{
uniquefy_by_variance(mat, map, dup_col = NULL)
}
\arguments{
\item{mat}{expression matrix}

\item{map}{gene map}

\item{dup_col}{col with potential duplicates}
}
\value{
filtered expression matrix
}
\description{
Uniquefies an matrix by the row variance. This
ranks variance for each as specifified in the
map data.frame. The row.names of the expression
matrix and map need to have at least some in common.
Function first identifies which row.names are shared
and then subsets.
Idea for using max var to de-duplicate probes came
form here: https://www.biostars.org/p/51756/#51875
}
