% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{venn_maker}
\alias{venn_maker}
\title{Venn maker}
\usage{
venn_maker(set_list, tit = "")
}
\arguments{
\item{set_list}{named vector list}

\item{tit}{plot title}
}
\value{

}
\description{
Venn maker
}
\keyword{internal}
