% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_wrappers.R
\name{voom_sva_wrapper}
\alias{voom_sva_wrapper}
\title{Runs voom to sva pipeline}
\usage{
voom_sva_wrapper(dge, mod, mod0)
}
\arguments{
\item{dge}{DGE object}

\item{mod}{model.matrix with variables of interest}

\item{mod0}{model.matrix without variables of interest}
}
\value{
voom object
}
\description{
This runs the voom to sva pipeline then voom
again pipeline. Returns a voom transformed
object fitted with the surrogate variables
}
