% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{color_dna_sequence}
\alias{color_dna_sequence}
\title{Colors a vector of DNA/peptide sequences via html}
\usage{
color_dna_sequence(sequences)
}
\arguments{
\item{sequences}{vector of DNA or peptide sequences}
}
\value{

}
\description{
Colors a vector of DNA/peptide sequences via html
}
\examples{
library(knitr)

# Example 1: DNA sequences with A, C, T, G
dna_sequences <- c("ACTG", "GCTA", "TACG")
colored_dna_sequences <- sapply(dna_sequences, color_dna_sequence)

# Creating a table with kable for DNA sequences
dna_data <- data.frame(
  DNA_Sequence = dna_sequences,
  Colored_DNA_Sequence = colored_dna_sequences
)
kable(dna_data, escape = FALSE)  # escape = FALSE allows rendering of HTML

# Example 2: Peptide sequences
peptide_sequences <- c("MKTAY", "LLGTP", "IVAGL")
colored_peptide_sequences <- sapply(peptide_sequences, color_dna_sequence)

# Creating a table with kable for peptide sequences
peptide_data <- data.frame(
  Peptide_Sequence = peptide_sequences,
  Colored_Peptide_Sequence = colored_peptide_sequences
)
kable(peptide_data, escape = FALSE)  # escape = FALSE allows rendering of HTML
}
